Add a static java waterfall client
